refactor: detail modal 컴포넌트에서 커스텀 훅 useDetailData로 api 로직 분리
